audio: address review comments on the finished player's position
Move the position tests from position_test.go to audio_test.go, remove
the test-name comments, use for-range and var for the loops, and use
defer for unlocking in the dummy player's Seek.
